add support for a .terraflow.yaml configuration file for setting default values and move options funtionality into its own file to remove circular dependencies
